+++ apps/bbsync/tests/conftest.py
@@ -1,9 +1,5 @@
-import uuid
-
 import pytest
-from django.conf import settings
 
-from osidb.core import generate_acls
 from osidb.tests.factories import FlawFactory
 
 
@@ -20,13 +16,3 @@ def test_flaw():
 @pytest.fixture
 def sentinel_err_message():
     return "400 Client Error: Bad Request"
-
-
-@pytest.fixture
-def internal_read_groups():
-    return [uuid.UUID(acl) for acl in generate_acls([settings.INTERNAL_READ_GROUP])]
-
-
-@pytest.fixture
-def internal_write_groups():
-    return [uuid.UUID(acl) for acl in generate_acls([settings.INTERNAL_WRITE_GROUP])]
