+++ osidb/tests/conftest.py
@@ -1,12 +1,9 @@
 import os
-import uuid
 from datetime import datetime, timezone
 from unittest.mock import MagicMock
 
 import pytest
-from django.conf import settings
 
-from osidb.core import generate_acls
 from osidb.helpers import get_env
 from osidb.integrations import IntegrationRepository, IntegrationSettings
 from osidb.models import FlawSource
@@ -57,36 +54,6 @@ def good_cve_id():
 @pytest.fixture
 def good_cve_id2():
     return "CVE-2021-999999"
-
-
-@pytest.fixture
-def public_read_groups():
-    return [uuid.UUID(acl) for acl in generate_acls(settings.PUBLIC_READ_GROUPS)]
-
-
-@pytest.fixture
-def embargoed_read_groups():
-    return [uuid.UUID(acl) for acl in generate_acls([settings.EMBARGO_READ_GROUP])]
-
-
-@pytest.fixture
-def internal_read_groups():
-    return [uuid.UUID(acl) for acl in generate_acls([settings.INTERNAL_READ_GROUP])]
-
-
-@pytest.fixture
-def public_write_groups():
-    return [uuid.UUID(acl) for acl in generate_acls([settings.PUBLIC_WRITE_GROUP])]
-
-
-@pytest.fixture
-def embargoed_write_groups():
-    return [uuid.UUID(acl) for acl in generate_acls([settings.EMBARGO_WRITE_GROUP])]
-
-
-@pytest.fixture
-def internal_write_groups():
-    return [uuid.UUID(acl) for acl in generate_acls([settings.INTERNAL_WRITE_GROUP])]
 
 
 @pytest.fixture
